Fix issues with serializing arrays of enum classes
Now that .mojom generates C++11 style enum classes for its enum types,
some of the array_serialization code needs to be tweaked.  In general,
the template specialization of ArraySerializer which handles PODs
shoud work with the addition of a cast, but simply adding the cast
could be a little bit dangerous.  Not only would it allow an enum
class to be stored as a specified storage type, it would also allow
other non-implicitly convertable types to be stored as each other
without warning.

So, in addition to adding the cast, also conditionalize the template
specializtion a bit more in order to avoid this potential future
hazard.  Specifically...

If the type to serialize is E, and the storage type is F...
+ Allow the template specialization to instantiate if E is already
  implicitly convertable to F.
Or...
+ Allow the template specialization to instantiate if E is is a scoped
  enumeration (aka, enum class), and F is exactly equal to E's
  underlying data type.

Also, add tests to exercise these cases.
Both array serialization and map serialization can trigger this (since
maps use array serialization internally), so add tests for..
+ Arrays of scoped enums
+ Maps of int values with scoped enum keys.
+ Maps of scoped enum values with int keys.

BUG=#468
R=viettrungluu@chromium.org

Review URL: https://codereview.chromium.org/1399673002 .