go.py: Use absolute paths to refer to files in a directory listing.
This fixes a bug I introduced in https://codereview.chromium.org/1384243002/ when I refactored go.py so that it is no longer the case that |build_dir| is the current working directory when some of the logic is executed. I thought I had fixed up all the cases where this mattered but I missed one.

I believe this was not caught on the bots because out directory is not purged.

R=viettrungluu@chromium.org

Review URL: https://codereview.chromium.org/1395583005 .