Pass src_root to InvokeGo from run_pure_go_tests.py
By passing src_root to InvokeGo() the Go code in one package is able to reference the Go code in a different package specified relative to src_root. This was not necessary for the lexer tests but it is necessary for the parser tests.

Also: make sure to use the absolute path to the go tool since we are changing directories before invoking it.

R=viettrungluu@chromium.org, azani

Review URL: https://codereview.chromium.org/1402793003 .