Produce system_thunks archive library and upload to cloud storage
This produces libsystem_thunks.a and uploads it to the Mojo cloud
storage bucket along with other artifacts. The file ends up at:
gs://mojo/system_thunks/{android-arm,linux-x64}/{VERISON}/libsystem_thunks.a

This library is useful when linking a Go (or potentially other language)
program that needs definitions of the system thunks setter but doesn't
want to bother with configuring a C compiler.

R=viettrungluu@chromium.org

Review URL: https://codereview.chromium.org/1403013002 .