Refactor mojo_benchmark to get rid of warm/cold start special cases.
This patch extracts a pre-processing step in mojo_benchmark that
rewrites the given list of benchmarks to run, yielding two benchmarks
corresponding to the warm start case and to the cold start case for each
benchmark defined in the original list.

This simplifies the logic afterwards which does not need to reason about
cold start vs warm start cases. This will also allow us to easily add
other variants in the future.

R=viettrungluu@chromium.org

Review URL: https://codereview.chromium.org/1400003005 .