Move observatory bundling rules to separate BUILD.gn file
Flutter bundles the observatory in the same way that the Dart content
handler does but it doesn't need the rest of the things in
//mojo/dart/embedder/BUILD.gn. Currently it copies the whole file and
then deletes some parts just to use the observatory rules which is a bit
silly. This moves the observatory rules into their own build file that
Flutter can reuse directly.  Soon the observatory resources won't need
to be bundled at all, at which point this whole thing can be deleted.

R=johnmccutchan@google.com

Review URL: https://codereview.chromium.org/1418473003 .