Remove many //base dependencies from //mojo/dart/embedder
This drops many of the //base dependencies for the C++ used to expose
Mojo primitives through the Dart VM so that it can be easier to share
with the Dart content handler and the Flutter engine.

The code currently also contains several policy decisions (such as what
goes in the Dart isolate storage slot) which will have to be dealt with
to make the code completely shared.

R=abarth@chromium.org

Review URL: https://codereview.chromium.org/1415463003 .