Improves the representation of values in Mojom.
This CL updates the representation of values in Mojom based on a better understanding obtained by working on the Go version of the parser. The main change is that we now have value keys for enum values as well as user-defined constants. Additionally some terminology was modified. We were using the word "constant" in places
where the word "value" makes more sense.

- Replace  map<string, DeclaredConstant> resolved_constants;
  with     map<string, UserDefinedValue> resolved_values;
  in MojomFileGraph.

- Add support for the "default" keyword as the default value of a struct field.

- Replace |ConstantOccurrence| with |Value|.

- Replace |ConstantValue| with |LiteralValue|.

- Replace |ConstantReference| with |UserValueReference|.

- Simplify representation of enum values.

R=azani@chromium.org, azani

Review URL: https://codereview.chromium.org/1417473002 .