Simplify mojo::internal::ArrayTraits.
Also, stop doing a whole bunch of unnecessary stuff in the move-only
case (some of it was necessary to handle move-only types, pre-C++11;
some of it -- like the alignment stuff -- was of rather dubious
necessity and correctness).

Probably it can/should be improved even more. E.g., PushBack() should be
reworked (into two overloaded versions, a const T& one and a T&& one).
That leaves only Clone(), which is a bit more tricky.

R=jamesr@chromium.org

Review URL: https://codereview.chromium.org/1417843002 .