Wrap download_from_google_storage.py in a cwd-independent script
gclient runhooks runs all of the hooks with a working directory set to
the location of the .gclient file (i.e. parent of src/) and
download_from_google_storage.py relies on this to find the sha1. This
wraps download_from_google_storage.py in a script that works from anywhere.

R=viettrungluu@chromium.org

Review URL: https://codereview.chromium.org/1410783004 .