Dart: Don't send unnecessary closes to the handle watcher
When the handle watcher tells us that the PEER_CLOSED signal is
satisfied for a handle, the handle watcher has likely already
closed and pruned the handle. Therefore, we only need to tear
down the local state, and don't need to send a close command
to the handle watcher.

BUG=
R=johnmccutchan@google.com

Review URL: https://codereview.chromium.org/1419233006 .