Dart: Proxy errors are routed to a future.
If a Proxy p encountes an error. The error is routed to p.errorFuture.
Proxies should no longer throw exceptions unless there is an internal
error. In case of an error during a Proxy call, the response future
will never complete.

This is a breaking change.

fixes #496

R=johnmccutchan@google.com

Review URL: https://codereview.chromium.org/1410373004 .