EDK: Add a simple CommandLine class.
(Currently, it resides in the EDK, but maybe we'll use it elsewhere
eventually too.)

It's similar to Chromium's base::CommandLine, in that it doesn't require
registration of options, etc. (It's up to the user to validate these
things, if they want to.) But it's not mutable.

It also doesn't provide a global singleton. If someone wants/needs that,
they should maintain that separately.

See the file-level comment in command_line.h for details about how it
processes options. (I call them "options": "flags" to me implies things
that are binary rather than things that may take values, while
"switches" is inconvenient since its singular is a keyword.)

R=jamesr@chromium.org

Review URL: https://codereview.chromium.org/1407863005 .