Expose JSON-encoded wire blessings in principal.User objects
This CL changes the User identity object in the Principal service
to directly exposes the JSON-encoded wire blessings obtained from
the Vanadium identity provider. These blessings are required for
the Mojo Syncbase service. Additionally, it gets rid of the Blessings
type defined in the principal service as that is no longer used.

R=gauthamt@google.com

Review URL: https://codereview.chromium.org/1414113011 .