Normalize names of rtt benchmarks.
This patch normalizes the names of the rtt benchmarks to match the style
of other benchmarks (in mojo/tools/data/rtt_benchmarks). This is in
preparation to enabling rtt benchmarks results to be uploaded to the
perf dashboard, where benchmarks names are used to label metrics.

R=qsr@chromium.org

Review URL: https://codereview.chromium.org/1409053010 .