Remove *.mojom.dart from //{,mojo/public/dart/}.gitignore, stop copying during build
We shouldn't have any mojom.dart files in the tree right now. When we do
we do not want to be .gitignore'ing them. We also should not be copying files from
the build directory into the source directory during a build. In practice this
seemed to only be copying files into //mojo/public/dart/mojo/sdk_ext/lib/ for
reasons not entirely clear to me.

R=johnmccutchan@google.com

Review URL: https://codereview.chromium.org/1421553017 .