Use an absolute path as the canonical name of a mojom file imported from another mojom file.
This fixes a bug in parse_driver.go in which when an imported file is found relative to a path in the import search path, then that imported file had its canonical name set to a relative path instead of an absolute path. We also add a test for this functionality.

R=azani@chromium.org, azani
BUG=#461

Review URL: https://codereview.chromium.org/1450593003 .