Add an ApplicationConnector interface, etc.
Also do a bit of (minimal) cleanup of ApplicationImpl.

From the Shell interface, you can mint ApplicationConnectors, which
allows you to connect to other applications. This is useful if you want
to connect to applications from other threads or other environments
(e.g., inside a VM).

TODO(vtl): Maybe add a Duplicate() to ApplicationConnector. The reason
we don't just add a Duplicate() to Shell is that Shell is associated
with low-level stuff (e.g., process lifetime), and we may want to add
more low-level process-y stuff (which shouldn't be delegated) to it in
the future.

R=jamesr@chromium.org

Review URL: https://codereview.chromium.org/1435153003 .