New Mojom Parser: Represent integer literals using the smallest possible integer type
R=azani@chromium.org, azani
BUG=#461

Prior to this patch the Mojom parser would store all integer literals in an int64. There are two
problems with this:
(a) We are unable to parse values larger than max(int64) but smaller than max(uint64)
(b) We are not representing any information about the size of the integer that would be useful
later for checking the validity of uses of the integer.

In this patch we change the parser to store integer literals in a variable whose type is the first
one from the following list in which the integer will fit:
int8, uint8, int16, uint16, int32, uint32, int64, uint64.

Note that in this patch we do not use the information about the size of the integer in any way.
See https://github.com/domokit/mojo/issues/539.

Summary of file changes:
- parsing.go
  = Update interface and logic in parsePositiveIntegerLiteral() to return information
    about the smallest size integer the parsed value will fit into.
  = Update parseNumberLiteral() to use the value returned from parsePositiveIntegerLiteral()
    in order to create a literal of the smallest possible type.
- types.go - Added Make*LiteralValue() functions for the 7  other integer types.
- scopes.go - I fixed an unrelated bug that bit me while writing my tests.

Review URL: https://codereview.chromium.org/1453003002 .