EDK: Actually provide an abstraction for task runners.
Previously, the embedder had to inject a notion of task runner via a
header file, which was pretty dubious. Now, the embedder gets to
implement and provide objects implementing an interface
(PlatformTaskRunner). So the embedder can have multiple types of task
runners (and message loops, etc.) if it wants. More importantly, it
makes the build dependency structure sane: the EDK doesn't actually
depend on the embedder.

Included is an implementation of PlatformTaskRunner
(base_edk::PlatformTaskRunnerImpl) for use with //base (and
base::TaskRunner), in //mojo/edk/base_edk. (This may be moved elsewhere
eventually, but currently EDK tests need it and //base.)

Things have become and will remain a little messy until I add
abstractions for message loops, etc. (E.g., for now we have to construct
a base::MessageLoop, and then a base_edk::PlatformTaskRunnerImpl
separately.)

R=jamesr@chromium.org

Review URL: https://codereview.chromium.org/1453243004 .