User lower case 'true' and 'false' to represent Boolean literals in the Python representation.
BUG=#547
R=azani@chromium.org, azani

The old compiler implicitly used 'true' and 'false' to represent boolean values in the Python "module" representation. This is because 'true' and 'false' are the Mojom representations and the old compiler never coerced these into anything other than strings. The new compiler does coerce these into boolean values and so mojom_translator.py was yielding the string values 'True' and 'False'. This patch corrects that to 'true' and  'false' since that it was the code generators expect.

Review URL: https://codereview.chromium.org/1457943002 .