New Mojom Compiler: Three fixes to mojom_translator.py
BUG=#547
R=azani@chromium.org, azani

This patch fixes three issues I found in mojom_translator.py:

(1) When translating floats and doubles the str() function does not produce a string that accurately captures the floating point number. Now we use the repr() function.

(2) We should treat union field tags the same way we treat struct field ordinals. At the moment these numbers are not being computed in the Go parser and they are being computed by code invoked from the code generators. We should reverse this after we fully-switch to the new compiler. But for now we need to not populate those fields in mojom_translator with unset values from mojom_types.mojom. This was being done correctly for struct field ordinals but not for union field tags. This patch corrects that.

(3) Some of the code generators expect the type of field names to be non-unicode strings. We should correct that
after we fully switch over to the new compiler but for now we convert our field names to non-unicode strings.

Review URL: https://codereview.chromium.org/1484743003 .