New Mojom parser: Compute enum value integer values.
BUG=#547

This patch implements the method MojomDescriptor.ComputeEnumValueIntegers().

Currently in mojom_translator.py we are consuming the |intValue| field from a |MojomEnumValue| (from mojom_types.mojom). But up until now this field has not been set by the parser. This patch corrects that.

The algorithm in computeEnumValueIntegersForEnum() in computed_data.go has been copied from the analogous code in ResolveNumericEnumValues() in mojo/public/tools/bindings/pylib/generate/data.py.

As part of this patch we also cleaned up the validation of enum value initializers in the parsing and resolution phase.
We tidied some of that code and improved the error messages.

Also as part of this patch we improved the debug output of the serialized MojomFileGraph.

R=azani@chromium.org

Review URL: https://codereview.chromium.org/1480413002 .