New Mojom parser: Adds |ImportedFrom| field to MojomFile.
BUG=#570
R=azani@chromium.org, azani

In the struct MojomFile we add the field
  ImportedFrom *MojomFile.

This will point to the "first reason" why the MojomFile was accessed by the parser. That is, if the MojomFile is a top-level file then |ImportedFrom| will be nil but if the MojomFile is not a top-level file but rather it was imported from one or more other files then |ImportedFrom| will point to the first importing file encountered by the parser.

This information will be used to enhance compilation error messages by including information about a chain of imports that explains why the mojom file is being compiled. Note that without this field we already do produce this type of message, but only for errors encountered during the parsing phase. For errors encountered during a later phase, such as resolution, we have lost the information and so cannot efficiently produce the message. This field allows us to do so.

Review URL: https://codereview.chromium.org/1488113003 .