New Mojom Parser: Clean up  error message generation.
BUG=#570
R=azani@chromium.org, azani

This patch cleans up the code that generates parsing error messages. Note that this patch
only deals with errors that occur during the parsing phase, not the later phases such as resolution.
We will deal with that in a later patch.

- The final error string is now constructed in a central location: ParseError.Error()
in parser.go. Previously the error strings were constructed in multiple places including
all over the place within parsing.go and in parse_driver.go.

- There is a uniform syntax for the prefix of the error string:
<file-name>:<line-number>:<column-number>

- The ParseError struct now contains |token| and |fileContents| fields so that
we can add snippets in a later patch.

- The ParseError struct now contains a |file| field so that we can add import path
info in a later patch.

- The helper function parseError() was introduced to cleanup error reporting in parsing.go.

- In addition to reporting ParseErrors, the parser also reports DuplicateNameErrors that are generated
in the Mojom layer. Previously these DuplicateNameErrors were being wrapped in a ParseError but that
didn't really make sense because the DuplicateNameErrors are already formed with their own prefixes
of the form <file-name>:<line-number>:<column-number>. We changed the field |err| in the Parser struct
to be of type |error| instead of type |ParseError| so that it can hold a ParseError or a DuplicateNameError.

Review URL: https://codereview.chromium.org/1489223002 .