New Mojom Parser: Finish implementing UserValueRef.validateAfterResolution()
This patch completes the logic for validating assignments. It takes care of the case that the right-hand-side of an assignment is a reference to a constant or an enum or a built-in-float-constant. In this case, after the value reference has been resolved, we validate the resolved value against the type of the variable to which it is being assigned. The main new logic is in UserValueRef.validateAfterResolution() in types.go. The methods TypeRef.IsAssignmentCompatible() and TypeRef.ResolvedTypeName() have been added to mojom_types.go to facilitate the implementation.

Some other miscellaneous changes:

- Introduced |AssigneeSpec| to hold both the name and the type of the left-hand-side of an assignment. Previously we were only storing the type and not the name but we need the name for an error message.

- Renamed |variableAssignement| to |literalAssignment| to represent the fact that it only deals with the case where the right-hand-side of a variable assignment is a literal. This is not the case we are dealing with in this patch. That case was already handled.

- Renamed |SimpleTypeFloat_INFINITY| to |FloatInfinity| and similarly with the other built-in-float-constants. The name
|SimpleTypeFloat_INFINITY|was due to an erroneous global search-and-replace in some earlier patch.

- Fixed up miscellaneous comments.

- Added a new test: resolution_test.go

BUG=#461
R=azani@chromium.org, azani

Review URL: https://codereview.chromium.org/1497373002 .