EDK: Remove mojo::platform::MessageLoopForIO and add PlatformHandleWatcher.
Instead, factories for "MessageLoopForIOs" should produce a MessageLoop
and (a pointer to) a PlatformHandleWatcher (which may be implemented by
the same object).

This is so when I make a factory for "I/O threads", it'll produce a
"Thread" object, a TaskRunner, and a PlatformHandleWatcher.

R=vardhan@google.com

Review URL: https://codereview.chromium.org/1507903006 .