Mojo C++ bindings: plumb validation errors down the stack, and add some more informative error messages.
There are still various validation errors that don't have an error
message, and this will be covered in a following CL.

* a MOJO_INTERNAL_DEBUG_SET_ERROR_MSG(<ptr to string>) macro that lets
  you format and set a string on debug. (in optimized NDEBUG builds, these
  strings are omitted all together).
* Plumb all the |Validate*| and validation-related functions to return
  a ValidationError instead of ReportValidationError.
* Plumb the Validate(..) methods generated for structs and unions (and
  also arrays & maps, which are not generated) to return a
  ValidationError and set a message.

Currently, incoming messages are validated by a MessageReceiver (part
of a FilterChain), which only pushes a Message along the FilterChain
without a way to plumb ValidationErrors. For this reason, we keep around
our ReportValidationError(), and the ValidationErrorObserver used for
unit tests.

R=viettrungluu@chromium.org
BUG=#520

Review URL: https://codereview.chromium.org/1496183002 .