Rename samples_per_frame back to channels.
This got switched from channels to samples_per_frame during the
original review, but since then there seem to have been some second
thoughts on the subject.  Move this back to channels to mesh a bit
better with some of Dale's work which is currently in flight.

R=jeffbrown@google.com, dalesat@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1492543002 .