New Mojom parser: Restrict the uses of the 'default' keyword.
BUG=#461

The new Mojom parser was mistakenly allowing the 'default' keyword to be used more generally than it is supported by the backends. This change fixes that. The 'default' keyword now may only be used as the default value of a struct field whose type is a struct.

Also we fix the error messages that complain about the mis-use of the default keyword.

R=azani@chromium.org

Review URL: https://codereview.chromium.org/1517093002 .