Remove -Wno-tautological-compare from fusl build
For unsigned types, x < 0 comparisons are removed.

For enum tyes, x < 0 comparisions are replaced with (int)x < 0. C99 does
not specify the type of an enum. For e.g. ns_sect which has only
nonnegative values, it could be either signed or unsigned. In either
case, the cast to int should be safe.

R=viettrungluu@chromium.org

Review URL: https://codereview.chromium.org/1578383002 .