Improve Ganesh helpers.
Properly reset and flush the Ganesh context when needed.

When intermixing Ganesh rendering with other GL operations, the
Ganesh renderer's cached GPU state can get out of sync with the
actual state of the GL context, resulting in broken behavior.

To fix this problem, reset the Ganesh context when entering its
scope and flush it when exiting its scope.  This operation has a
small performance cost but only when entering and exiting the scope
which should be infrequent (ideally no more than once per frame).

Added a helper for creating SkImages from a texture and updated
the GaneshTextureSurface to use the new GrGLTextureInfo type
which was introduced in the latest Skia roll.

Moved helpers into the mojo::skia namespace.

R=abarth@google.com

Review URL: https://codereview.chromium.org/1534033002 .