Add a multi-threaded rendering example.
Add a new example to test multi-threaded rendering support to ensure
that the Mozart APIs are not overly thread-hostile.

The Noodles app draws Lissajous figures.  The main thread handles
events and submits SkPictures to the rasterizer thread to be
drawn and composited.

Writing this example provided useful feedback into the design of the
UI helper library and the Scene interface itself, particularly the
separation of event handling from content publishing.

BUG=
R=abarth@google.com

Review URL: https://codereview.chromium.org/1558813002 .