Dart: Roll runtime forward
This roll will have a negative impact on startup performance due to
the need to eagerly compute a Uri for the script. See the comment
in builtin.dart.

Also, CreateIsolateCallback is now always passed a script_uri, even
for an Isolate.spawn, so that code is re-jiggered a bit as well.

R=johnmccutchan@google.com

Review URL: https://codereview.chromium.org/1640933003 .