[fusl] Add llvm's lit tool to third_party
This will be used for running libcxx's tests on top of fusl.

This code is a directory in upstream llvm. It is also a PyPi
package: https://pypi.python.org/pypi/lit. It's not particularly easy to
consume in domokit as either of those things, though. Since it doesn't
change upstream much at all, let's just check in a copy.

This was taken from llvm revision 250162. It is exactly the contents of
the //utils/lit directory, plus the LICENSE.TXT file.

R=phosek@chromium.org, jamesr@chromium.org

Review URL: https://codereview.chromium.org/1663053003 .