Mojom parser: Introduces UserDefinedType.FindReachableTypes().
This method returns the set of type keys of all types that are reachable from a given type in the type graph. The method is implemented as a simple depth-first-search.

The intended use case is to invoke this method only on the top-level interface of a service during the serialization phase.

BUG=#448
R=azani@chromium.org

Review URL: https://codereview.chromium.org/1661653006 .